requirements.txt
contextily==1.6.2
dropblock==0.3.0
einops==0.8.1
exifread==3.4.0
geopandas==0.10.2
ipython==8.12.3
kitti360Scripts==1.0.0
lxml==5.3.0
matplotlib==3.5.3
numba==0.60.0
numpy==1.23.0
omegaconf==2.3.0
opencv_python==4.10.0.84
pandas==1.5.3
Pillow==11.3.0
plotly==5.13.1
pykitti==0.3.1
pyproj==3.7.0
pyquaternion==0.9.9
pytorch_lightning==2.5.0.post0
PyYAML==6.0.2
Requests==2.32.5
Rtree==1.3.0
scikit_learn==1.2.2
scipy==1.16.1
sequential==1.0.0
setuptools==67.6.0
Shapely==1.8.5.post1
tabulate==0.9.0
torch==2.3.1
torch_scatter==2.1.2+pt23cu121
torchvision==0.18.1
tqdm==4.65.2
urllib3==1.26.20
